prepare automated testing for github action
