+++ matchmaker/io/midi.py
@@ -13,13 +13,13 @@ from mido.ports import BaseInput as MidiInputPort
 
 from matchmaker.features.midi import PitchIOIProcessor
 from matchmaker.io.mediator import CeusMediator
+from matchmaker.utils.midi import get_available_midi_port
 from matchmaker.utils.misc import RECVQueue
 from matchmaker.utils.processor import Processor
 from matchmaker.utils.stream import Stream
 from matchmaker.utils.symbolic import (
     Buffer,
     framed_midi_messages_from_performance,
-    get_available_midi_port,
     midi_messages_from_performance,
 )
